requirements.txt
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==2.2.10
django-crispy-forms==1.9.1
djangorestframework==3.11.0
get==0.0.0
httpie==2.2.0
idna==2.10
post==0.0.0
Pygments==2.6.1
pytz==2020.1
query-string==0.0.0
request==0.0.0
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
wincertstore==0.2
